requirements.txt
flask
flask-pymongo
python-dotenv
sentence-transformers
faiss-cpu
numpy
groq
pymongo